Use process.env to retrieve $EDITOR
The $ notation to fetch environment variables does not work on
windows. Instead, get the environment variable from NodeJS using
process.env.EDITOR

The "spawn" command by default does not use the systems environment
variables. The variable PATH is not included, causing some paths
not to be resolved on windows. Passed process.env through to spawn
to make sure paths are resolved with PATH on windows.